requirements.txt
flask 
numpy 
pillow
requests
cors
flask-cors